Refactor Hadoop connectors
