+++ hazelcast-jet-hadoop/src/main/java/com/hazelcast/jet/impl/connector/hadoop/WritableSerializerHooks.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.jet.connector.hadoop;
+package com.hazelcast.jet.impl.connector.hadoop;
 
 import com.hazelcast.logging.ILogger;
 import com.hazelcast.logging.Logger;
