+++ hazelcast-jet-hadoop/src/main/java/com/hazelcast/jet/processors/package-info.java
@@ -15,6 +15,6 @@
  */
 
 /**
- * HDFS read/write support for Hazelcast Jet
+ * HDFS read/write support for Hazelcast Jet.
  */
-package com.hazelcast.jet.connector.hadoop;
+package com.hazelcast.jet.processors;
