+++ hazelcast-jet-hadoop/src/test/java/com/hazelcast/jet/impl/connector/hadoop/WriteHdfsPTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.jet.connector.hadoop;
+package com.hazelcast.jet.impl.connector.hadoop;
 
 import com.hazelcast.core.IList;
 import com.hazelcast.jet.DAG;
@@ -52,8 +52,8 @@ import java.util.stream.IntStream;
 import static com.hazelcast.jet.Edge.between;
 import static com.hazelcast.jet.processor.Sources.readMap;
 import static com.hazelcast.jet.processor.Sinks.writeList;
-import static com.hazelcast.jet.connector.hadoop.ReadHdfsP.readHdfs;
-import static com.hazelcast.jet.connector.hadoop.WriteHdfsP.writeHdfs;
+import static com.hazelcast.jet.processors.HdfsProcessors.readHdfs;
+import static com.hazelcast.jet.processors.HdfsProcessors.writeHdfs;
 import static java.util.stream.Collectors.toMap;
 import static org.junit.Assert.assertEquals;
